fix: refactor map tracking control
- move control to map folder
- remove unused and constant props
- remove duplicate icons
- remove hardcoded icon colors
